Added Animations And input Changes
added basic movement and jumping animations to the character and changed the collider parent to the 3d mesh hip bone(Root bone) so collider can follow anywhere the player moves or animates. And changes the input and camera rotation and made it less lines with unity math functions in Movement scripts.